move facet storage to transaction module
